refactor: presentational로 분리한다
